refactoring: move DeliveryAddressDataApiFactory to DeliveryAddressMutation namespace
