+++ project-base/app/src/FrontendApi/Mutation/Customer/DeliveryAddressMutation.php
@@ -11,7 +11,7 @@ use Shopsys\FrontendApiBundle\Model\Mutation\Customer\DeliveryAddress\DeliveryAd
  * @property \App\Model\Customer\User\CurrentCustomerUser $currentCustomerUser
  * @property \App\Model\Customer\User\CustomerUserUpdateDataFactory $customerUserUpdateDataFactory
  * @property \App\Model\Customer\User\CustomerUserFacade $customerUserFacade
- * @method __construct(\App\Model\Customer\DeliveryAddressFacade $deliveryAddressFacade, \App\Model\Customer\User\CurrentCustomerUser $currentCustomerUser, \App\Model\Customer\DeliveryAddressDataFactory $deliveryAddressDataFactory, \App\Model\Customer\User\CustomerUserUpdateDataFactory $customerUserUpdateDataFactory, \App\Model\Customer\User\CustomerUserFacade $customerUserFacade)
+ * @method __construct(\Symfony\Component\Security\Core\Authentication\Token\Storage\TokenStorageInterface $tokenStorage, \App\Model\Customer\DeliveryAddressFacade $deliveryAddressFacade, \App\Model\Customer\User\CustomerUserUpdateDataFactory $customerUserUpdateDataFactory, \App\Model\Customer\User\CustomerUserFacade $customerUserFacade, \Shopsys\FrontendApiBundle\Model\Mutation\Customer\DeliveryAddress\DeliveryAddressDataApiFactory $deliveryAddressDataApiFactory)
  * @method \App\Model\Customer\DeliveryAddress[] deleteDeliveryAddressMutation(\Overblog\GraphQLBundle\Definition\Argument $argument)
  * @method \App\Model\Customer\DeliveryAddress[] editDeliveryAddressMutation(\Overblog\GraphQLBundle\Definition\Argument $argument)
  * @method \App\Model\Customer\User\CustomerUser setDefaultDeliveryAddressMutation(string $deliveryAddressUuid)
